Put modules of examples in separate files
