+++ examples/basics/main.rs
@@ -1,5 +1,12 @@
 use gtk::prelude::*;
 
+fn main() {
+    let application =
+        gtk::Application::new(Some("com.github.gtk-rs.examples.basic"), Default::default());
+    application.connect_activate(build_ui);
+    application.run();
+}
+
 fn build_ui(application: &gtk::Application) {
     let window = gtk::ApplicationWindow::new(application);
 
@@ -11,11 +18,4 @@ fn build_ui(application: &gtk::Application) {
     window.set_child(Some(&button));
 
     window.show();
-}
-
-fn main() {
-    let application =
-        gtk::Application::new(Some("com.github.gtk-rs.examples.basic"), Default::default());
-    application.connect_activate(build_ui);
-    application.run();
 }
