+++ examples/builder_pattern/main.rs
@@ -1,6 +1,15 @@
 use gtk::prelude::*;
 use gtk::{Align, Application, ApplicationWindowBuilder, ButtonBuilder};
 
+fn main() {
+    let application = Application::new(
+        Some("com.github.gtk-rs.examples.builder_pattern"),
+        Default::default(),
+    );
+    application.connect_activate(build_ui);
+    application.run();
+}
+
 fn build_ui(application: &Application) {
     let window = ApplicationWindowBuilder::new()
         .application(application)
@@ -22,11 +31,4 @@ fn build_ui(application: &Application) {
     window.set_child(Some(&button));
 
     window.show();
-}
-
-fn main() {
-    let application =
-        Application::new(Some("com.github.gtk-rs.examples.basic"), Default::default());
-    application.connect_activate(build_ui);
-    application.run();
 }
