+++ examples/composite_template/ex_application_window/ex_application_window.ui
