+++ examples/composite_template/ex_application_window/mod.rs
@@ -1,0 +1,24 @@
+mod imp;
+
+use gtk::subclass::prelude::*;
+use gtk::{gio, glib};
+
+glib::wrapper! {
+    pub struct ExApplicationWindow(ObjectSubclass<imp::ExApplicationWindow>)
+        @extends gtk::Widget, gtk::Window, gtk::ApplicationWindow, @implements gio::ActionMap, gio::ActionGroup;
+}
+
+impl ExApplicationWindow {
+    pub fn new<P: glib::IsA<gtk::Application>>(app: &P) -> Self {
+        glib::Object::new(&[("application", app)]).expect("Failed to create ExApplicationWindow")
+    }
+
+    pub fn init_label(&self) {
+        // To access fields such as template children, you must get
+        // the private struct.
+        let self_ = imp::ExApplicationWindow::from_instance(self);
+        self_
+            .subtitle
+            .set_text("This is an example window made using composite templates");
+    }
+}
