+++ examples/composite_template/ex_menu_button/ex_menu_button.ui
