+++ examples/composite_template/ex_menu_button/mod.rs
@@ -1,0 +1,7 @@
+mod imp;
+
+use gtk::glib;
+
+glib::wrapper! {
+    pub struct ExMenuButton(ObjectSubclass<imp::ExMenuButton>) @extends gtk::Widget;
+}
