+++ examples/css/style.css
@@ -1,0 +1,25 @@
+entry.entry1 {
+    background: linear-gradient(to right, #f00, #0f0);
+    color: blue;
+    font-weight: bold;
+}
+
+button {
+    /* If we don't put it, the yellow background won't be visible */
+    background-image: none;
+}
+
+button.button1:hover {
+    transition: 500ms;
+    color: red;
+    background-color: yellow;
+}
+
+combobox button.combo box {
+    padding: 5px;
+}
+combobox box arrow {
+    -gtk-icon-source: -gtk-icontheme('pan-down-symbolic');
+    border-left: 5px solid transparent;
+    border-right: 5px solid transparent;
+    border-top: 5px solid black;
