+++ examples/custom_editable/style.css
@@ -1,0 +1,23 @@
+tag {
+    margin: 4px 0px;
+    padding: 4px;
+    border-radius: 4px;
+    background: lightskyblue;
+  }
+  tag box {
+    border-spacing: 4px;
+  }
+  tag label,
+  tag image {
+    color: white;
+  }
+  tag button {
+    min-height: 0;
+    min-width: 0;
+    padding: 0;
+    border: 1px solid white;
+  }
+  
+  entry.tagged {
+    border-spacing: 4px;
+  }
