+++ examples/custom_layout_manager/custom_layout_child/mod.rs
@@ -1,0 +1,14 @@
+mod imp;
+
+use gtk::{gdk, glib};
+
+glib::wrapper! {
+    pub struct CustomLayoutChild(ObjectSubclass<imp::CustomLayoutChild>)
+        @extends gtk::Widget;
+}
+
+impl CustomLayoutChild {
+    pub fn new(color: gdk::RGBA) -> Self {
+        glib::Object::new(&[("color", &color)]).expect("Failed to create CustomLayoutChild")
+    }
+}
