+++ examples/custom_widget/ex_button/mod.rs
@@ -1,0 +1,20 @@
+mod imp;
+
+use gtk::glib;
+
+glib::wrapper! {
+    pub struct ExButton(ObjectSubclass<imp::ExButton>)
+        @extends gtk::Widget;
+}
+
+impl Default for ExButton {
+    fn default() -> Self {
+        Self::new()
+    }
+}
+
+impl ExButton {
+    pub fn new() -> Self {
+        glib::Object::new(&[]).expect("Failed to create Button")
+    }
+}
