+++ examples/video_player/video_player_window/video_player_window.ui
